+++ src/markdown-it-plugin.js
